Update memory-related methods (GetStringBuffer, AllocateMemoryPool, ReleaseMemoryPool) to use size_t instead of int for size and index parameters.
This improves type safety and aligns with standard library
conventions where size_t is the appropriate type for sizes and indices.

BREAKING CHANGE: Parameter types changed from int to size_t in GetStringBuffer,
AllocateMemoryPool, and ReleaseMemoryPool methods

+++ include/CKContext.h
@@ -162,7 +162,7 @@ public:
 
     //-------------------------------------------------------
     // Utils
-    CKSTRING GetStringBuffer(int size);
+    CKSTRING GetStringBuffer(size_t size);
     CKGUID GetSecureGuid();
     CKDWORD GetStartOptions();
     WIN_HANDLE GetMainWindow();
@@ -263,8 +263,8 @@ public:
     void BuildSortedLists();
     void DeferredDestroyObjects(CK_ID *obj_ids, int Count, CKDependencies *Dependencies, CKDWORD Flags);
 
-    void *AllocateMemoryPool(int count, int &index);
-    void ReleaseMemoryPool(int index);
+    void *AllocateMemoryPool(size_t count, size_t &index);
+    void ReleaseMemoryPool(size_t index);
 
     CKContext(WIN_HANDLE iWin, int iRenderEngine, CKDWORD Flags);
     ~CKContext();
